pyproject.toml
[project]
name = "RAG-end-to-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langchain-postgres>=0.0.16",
    "langchain-upstage>=0.7.5",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-sdk>=0.2.9",
    "langsmith>=0.4.46",
    "markdownify>=1.2.2",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.13",
    "pymupdf>=1.26.6",
    "rank-bm25>=0.2.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "mypy>=1.13.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "fitz",
    "fitz.*",
    "markdownify",
    "markdownify.*",
    "langchain_teddynote",
    "langchain_teddynote.*",
    "pandas",
    "pandas.*",
]
ignore_missing_imports = true

[tool.setuptools.packages.find]
# Explicitly include the packages you want to distribute/install
include = ["scripts", "utils", "reranker", "evaluators"]
# (Optional) Explicitly exclude things to be safe
exclude = ["data", "dev", "images", "outputs", "faiss_index", "test"]

[tool.uv.workspace]
members = [
    ".",
]
